Fix bug on windows preventing display in WebGL.
This might have been caused by a difference of big-small endian. I think in my previous implementation, Unsigned int were magically transformed into floats, but that won't work when we switch endian. Anyway, that was incorrect, so this fix solved it.

+++ public/lib/texture/texture-atlas.js
@@ -14,7 +14,7 @@ class TextureAtlas {
 		this.spriteWidth = 0;
 		this.spriteHeight = 0;
 
-		this.tempMatrix = new Uint16Array([
+		this.tempMatrix = new Float32Array([
 			0, 0, 0, 0,
 			0, 0, 0, 0,
 			0, 0, 0, 0,
